+++ siopv2/src/client_metadata.rs
