--- src/token/mod.rs
@@ -1,2 +1,0 @@
-pub mod id_token;
-pub mod id_token_builder;
